refactor: improvement control centralization of repo TODO
