+++ src/app/features/todos/todos_repository/todos_repository.ts
@@ -1,7 +1,11 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Injectable, signal } from '@angular/core';
+import { ToastController, AlertController } from '@ionic/angular';
+import { TranslateService } from '@ngx-translate/core';
+import { lastValueFrom, Observable } from 'rxjs';
 import { TodosApi } from '../todosAPI/todos_api';
 import { Todo } from '../todosAPI/models/todo';
+
+export enum TodosViewFilter { all, activeOnly, completedOnly }
 
 /**
  * A repository that handles `todo` related requests.
@@ -10,7 +14,20 @@ import { Todo } from '../todosAPI/models/todo';
   providedIn: 'root'
 })
 export class TodosRepository {
-  constructor(private todosApi: TodosApi) {}
+  todosSignal = signal<Todo[]>([]);
+  filterSignal = signal<TodosViewFilter>(TodosViewFilter.all);
+  lastDeletedTodoSignal = signal<Todo | null>(null);
+  lastDeletedTodoIndexSignal = signal<number | null>(null);
+
+  todos = this.todosSignal.asReadonly();
+  lastDeletedTodo = this.lastDeletedTodoSignal.asReadonly();
+
+  constructor(
+    private todosApi: TodosApi,
+    private alertController: AlertController,
+    private toastController: ToastController,
+    private translateService: TranslateService
+  ) { }
 
   /**
    * Provides an Observable of all todos.
@@ -46,9 +63,99 @@ export class TodosRepository {
    *
    * If no todo with the given id exists, a TodoNotFoundException is thrown.
    */
-  deleteTodo(id: string): Promise<void> {
-    return this.todosApi.deleteTodo(id);
+  async deleteTodo(todo: Todo): Promise<boolean> {
+    if (await this.confirmDelete(todo)) {
+      const todos = this.todos();
+      const index = todos.findIndex(t => t.id === todo.id);
+
+      if (index !== -1) {
+        this.lastDeletedTodoSignal.set(todo);
+        this.lastDeletedTodoIndexSignal.set(index);
+        await this.todosApi.deleteTodo(todo.id);
+        this.removeTodoFromList(todo);
+        this.showUndoToast(todo);
+      }
+      return true;
+    } else {
+      return false;
+    }
   }
+
+  private removeTodoFromList(todoToRemove: Todo) {
+    this.todosSignal.update(todos =>
+      todos.filter(todo => todo.id !== todoToRemove.id)
+    );
+  }
+
+
+  async confirmDelete(todo: Todo): Promise<boolean> {
+    return new Promise<boolean>(async (resolve) => {
+      const alert = await this.alertController.create({
+        header: await lastValueFrom(this.translateService.get('TODO_LIST_ITEM.DELETE_CONFIRM_HEADER')),
+        message: await lastValueFrom(this.translateService.get('TODO_LIST_ITEM.DELETE_CONFIRM_MESSAGE', { title: todo.title })),
+        buttons: [
+          {
+            text: await lastValueFrom(this.translateService.get('COMMON.CANCEL')),
+            role: 'cancel',
+            handler: () => {
+              resolve(false);
+            }
+          },
+          {
+            text: await lastValueFrom(this.translateService.get('COMMON.DELETE')),
+            role: 'destructive',
+            handler: async () => {
+              resolve(true);
+            }
+          }
+        ]
+      });
+      alert.present();
+    });
+  }
+
+
+  private async showUndoToast(deletedTodo: Todo) {
+    const toast = await this.toastController.create({
+      message: await lastValueFrom(this.translateService.get('TODOS_OVERVIEW.TODO_DELETED_MESSAGE', { title: deletedTodo.title })),
+      duration: 4000,
+      position: 'bottom',
+      buttons: [
+        {
+          side: 'end',
+          text: await lastValueFrom(this.translateService.get('TODOS_OVERVIEW.UNDO_DELETION_BUTTON')),
+          handler: () => {
+            this.undoDeleteTodo();
+          }
+        }
+      ]
+    });
+    await toast.present();
+  }
+
+  private async undoDeleteTodo() {
+    const todo = this.lastDeletedTodo();
+    const index = this.lastDeletedTodoIndexSignal();
+
+    if (todo && index !== null) {
+      await this.saveTodoAt(todo, index);
+      this.lastDeletedTodoSignal.set(null);
+      this.lastDeletedTodoIndexSignal.set(null);
+      this.addTodoToListAt(todo, index);
+    }
+  }
+
+  private addTodoToListAt(todoToAdd: Todo, index: number) {
+    this.todosSignal.update(todos => {
+      if (!todos.some(t => t.id === todoToAdd.id)) {
+        const newTodos = [...todos];
+        newTodos.splice(index, 0, todoToAdd);
+        return newTodos;
+      }
+      return todos;
+    });
+  }
+
 
   /**
    * Deletes all completed todos.
@@ -67,4 +174,35 @@ export class TodosRepository {
   completeAll(isCompleted: boolean): Promise<number> {
     return this.todosApi.completeAll(isCompleted);
   }
+
+  async loadTodos() {
+    this.getTodos().subscribe(todos => {
+      this.todosSignal.set(todos);
+    });
+  }
+
+
+  setFilter(filter: TodosViewFilter) {
+    this.filterSignal.set(filter);
+  }
+
+  async toggleAll() {
+    const areAllCompleted = this.todos().every(todo => todo.isCompleted);
+    await this.completeAll(!areAllCompleted);
+    await this.loadTodos();
+  }
+
+  async toggleTodoCompletion(todo: Todo, isCompleted: boolean) {
+    const newTodo = todo.copyWith({ isCompleted });
+    await this.saveTodo(newTodo);
+    this.updateTodoInList(newTodo);
+  }
+
+
+  private updateTodoInList(updatedTodo: Todo) {
+    this.todosSignal.update(todos =>
+      todos.map(todo => todo.id === updatedTodo.id ? updatedTodo : todo)
+    );
+  }
+
 }
