+++ src/app/features/todos/ui/edit-todo/edit-todo.service.ts
@@ -1,8 +1,6 @@
 import { Injectable, signal, computed } from '@angular/core';
-import { AlertController } from '@ionic/angular';
 import { Todo } from '../../todosAPI/models/todo';
 import { TodosRepository } from '../../todos_repository/todos_repository';
-import { TranslateService } from '@ngx-translate/core';
 import { lastValueFrom } from 'rxjs';
 
 export enum EditTodoStatus { initial, loading, success, failure }
@@ -26,11 +24,7 @@ export class EditTodoService {
     [EditTodoStatus.loading, EditTodoStatus.success].includes(this.statusSignal())
   );
 
-  constructor(
-    private todosRepository: TodosRepository,
-    private alertController: AlertController,
-    private translateService: TranslateService
-  ) { }
+  constructor(private todosRepository: TodosRepository) { }
 
   initializeTodo(todo: Todo | null) {
     this.statusSignal.set(EditTodoStatus.initial);
@@ -69,33 +63,10 @@ export class EditTodoService {
     }
   }
 
-  async confirmDelete(todo: Todo) {
+  async deleteTodo(todo: Todo): Promise<boolean>{
     this.statusSignal.set(EditTodoStatus.loading);
-    return new Promise(async (resolve) => {
-      const alert = await this.alertController.create({
-        header: await lastValueFrom(this.translateService.get('TODO_LIST_ITEM.DELETE_CONFIRM_HEADER')),
-        message: await lastValueFrom(this.translateService.get('TODO_LIST_ITEM.DELETE_CONFIRM_MESSAGE', { title: todo.title })),
-        buttons: [
-          {
-            text: await lastValueFrom(this.translateService.get('COMMON.CANCEL')),
-            role: 'cancel',
-            handler: () => {
-              this.statusSignal.set(EditTodoStatus.initial);
-              resolve(false);
-            }
-          },
-          {
-            text: await lastValueFrom(this.translateService.get('COMMON.DELETE')),
-            role: 'destructive',
-            handler: () => {
-              this.todosRepository.deleteTodo(todo.id);
-              this.statusSignal.set(EditTodoStatus.success);
-              resolve(true);
-            }
-          }
-        ]
-      });
-      alert.present();
-    });
+    const result = this.todosRepository.deleteTodo(todo);
+    this.statusSignal.set(EditTodoStatus.success);
+    return result;
   }
 }
